[Feat] m2 하단 text 삽입
