+++ frontend/lib/screen/main2_screen.dart
@@ -1,7 +1,6 @@
 import 'package:dodo/components/m2_board.dart';
 import 'package:dodo/const/colors.dart';
 import 'package:flutter/material.dart';
-import 'package:table_calendar/table_calendar.dart';
 
 class main2Page extends StatefulWidget {
   const main2Page({super.key});
@@ -45,28 +44,9 @@ class _main2PageState extends State<main2Page> {
                 ),
               ),
               m2_board(),
-              TableCalendar(
-                onDaySelected: onDaySelected,
-                selectedDayPredicate: (day) {
-                  return isSameDay(selectedDate, day);
-                },
-                focusedDay: selectedDate,
-                firstDay: DateTime(2024),
-                lastDay: DateTime(2025),
-                headerStyle: HeaderStyle(
-                    titleCentered: true,
-                    formatButtonVisible: false,
-                    titleTextStyle: TextStyle(fontFamily: 'bm', fontSize: 20)),
-              ),
             ],
           )),
     );
-  }
-
-  void onDaySelected(DateTime selectedDate, DateTime focusedDate) {
-    setState(() {
-      this.selectedDate = selectedDate;
-    });
   }
 
   Widget _roomBtn() {
